Fixes the arrow keys and all other keys that generate E0/E1 scancode prefixes on AT keyboards by adding support for E0/E1 scancode prefixes to the dummy keyboard driver if CONFIG_INPUT=y.
Rationale: When using the new input layer (i.e. with a USB keyboard or
a custom input device), the input layer relies on kbd_translate() in
the low-level keyboard driver to convert from AT-style scancodes to
keycodes. If you don't have a PS/2 keyboard interface and don't compile
in the PS/2 keyboard driver, you have to enable the dummy keyboard
driver, which naively assumes that keycodes and scancodes are
interchangeable. This is correct if you do not have a keyboard, but
fails for prefixed scancodes if you do have a keyboard which uses the
new input layer.